requirements.txt
python-socketio
eventlet
flask